services/streaming_service/requirements.txt
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.3
pycodestyle==2.7.0
python-dotenv==0.17.1
python-ffmpeg-video-streaming==0.1.14
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
